use QHtsp as library in release build
Previous QHtsp was being build as library, but it wasn't being used. This was
because the MeeTV just added all the files of the project to it's own project.
From now on QHtsp will be used as a real library, which will be linked to, in
the release build. For the debug build it's still using all the files, and not
the library, as this eases debugging.